Changed a bunch of code
Round results now call in other methods and print out the string format of the number that they have chosen. Tried looping the game if the user agrees to restart the whole game. The formula for computer's random choice has been modified and changed (Might change again in the future) Fixed numbers corresponding on the chosen move.